refactor: update charts folds and files
